Move extract workflow implementation to own python file
